(From Sierra) Changed some int values to size_t.
Added an test to verify that every processor is involved in the
detachment and skin-update operations.

Changed the number of skin entities test to work correctly for all NX >= 3.

Modified the size of the fixture to allow this problem to run out to 500
processors and still have every processor involved in the detachment and
skin update operations.  In this new test case, every processor now has
5000 elements on it.

Build/Test Cases Summary
Enabled Packages: STK
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=6,notpassed=0 (3.09 min)
1) SERIAL_RELEASE => passed: passed=6,notpassed=0 (2.11 min)