Fix issues with the TDD script usage on s909348
I was previously using a variable which isn't pulled in by the TDD scripts to
determine whether or not we are on a release branch of dev branch. I have
rewritten it to be based off the information in TrilinosVersion.cmake and
specifically which repository branch we are on. This will work so long as the
development branch is always master, everything else is assumed to be a release
branch. Since this machine only handles development and release testing I think
this is reasonable.