(From Sierra) FRIO: Add mapping of state variable fields using the defined state variable name maps in lame
This commit extends the use of the field mapping capability to the lame state variables.
Instead of the state variables appearing as state_01, state_02, state_03, ..., state_20.  As an example, they will now appear as,

	state_BACK_STRESS_XX
	state_BACK_STRESS_YY
	state_BACK_STRESS_ZZ
	state_BACK_STRESS_XY
	state_BACK_STRESS_YZ
	state_BACK_STRESS_ZX
	state_DKAPP
	state_PHI
	state_PHIDOT
	state_THETA
	state_THETADOT
	state_A_XX
	state_A_YY
	state_A_ZZ
	state_A_XY
	state_A_YZ
	state_A_ZX
	state_EQPS
	state_EQPSDOT

The suffices are defined by the lame material model state variable aliases.  This should make
it easier for the user to interpret their output files and cause less
confusion on what is what.

If a material model does not define any aliases, or only defines a
partial set of aliases, then the unnamed variables will use the old _#
naming.

One issue with this is that it is easy to overrun the 32-character
name limit on variable names.  The IO will truncate the name, but this
can lead to duplicate names since the distinguising characters are
often at the end.  Judicious naming of the output variable can help
here.

Additional related changes:
IOSS: Eliminate case sensitivity in variable type registration
STRUMENTO: See if there is a field mapping before adding a field to output
FRIO: Do not do field mapping for restart output
FRIO: Use field mapping if defined for the field being output
FRIO: Add names for the last two state variables in BCJ model for use with the field suffix mapping