Changed configure.ac to handle all TPLs identically and as documented.
Umit discovered last week that he could not use autotools without
Scotch; the configuration would build and say Scotch was not requested,
but make tried to compile scotch files anyway.  The handling of TPLs
in configure.ac differed for ParMETIS, Scotch and PaToH.  I changed
configure.ac to require --with-<TPLname> to use the TPL; I removed the
assumption that the TPL should be used if the TPL's library or include
path were specified.  This approach is consistent with CMAKE and is now
consistent across all TPLs.

I also removed the DRUM TPL; it is no longer supported.

All the other files committed here were affected by the changes to
configure.ac during the bootstrap.