Added print() function with no args for debugging, cleaned up formatting
Here I added the function PL::print() that takes no arguments.  This allows
calling print() in a debugger like GDB without it crashing.  This was very
helpful in debugging a recent problem.  One consequence of this is that now
Teuchos::ParameterList has a mandatory dependence on
Teuchos::VerboseObjectBase in order to get at the default output stream
object.

I also cleaned up the formatting of the Teuchos_ParameterList.cpp file
according to the Thyra Coding and Documentation Guidelines.

The only unit test for this new function is another unit test.  Given the
simplistic nature of this code I think this is okay.