New stacktrace test, runtime control of stacktrace, added Teuchos_ENABLE_DEFAULT_STRACKTRACE
With this commit, now you can turns stracktracing support on and off at
runtime.  This will allow me to set Teuchos_ENABLE_DEFAULT_STRACKTRACE=OFF in
the checkin-test.py script but turn it on only in tests for stacktracing.  The
default mode for developers and users will be to have it turned on in a debug
build.

I also added a new test file for TestForException.hpp macros to test
stacktracing.