Changing to set Teuchos_ENABLE_STRACKTRACE=ON always but setting Teuchos_ENABLE_DEFUALT_STACKTRACE based on debug.
With this new behavior, the ability to turn stack tracing is always enabled if
the dependencies are met.  However, stacktracing is not turned on by default
at runtime unless a debug build is being done.  This will give better testing
but will not make the release testing or default use of Trilinos take any
longer.

Build/Test Cases Summary
Enabled Packages: Teuchos
0) MPI_DEBUG => passed: passed=62,notpassed=0 (0.18 min)
1) SERIAL_RELEASE => passed: passed=59,notpassed=0 (0.14 min)