Changing pretty log alias from using (%ar) to (%cr).
The old format was showing the dates as:

   [%ad] ($ar)

which was showing the same author date in two ways as [absolute author date]
(relative author date).  However, I decided that it would be useful to have
the relative date instead show the committer date which will be different if
the commit message is amended in anyway (e.g. through rebasing).

The new format for dates are:

   [%ad] ($cr)

I think that showing the reative committer date instead of the relative author
date gives more information in the short log.

I think this change.