Removing the grep for loc_segfault_callback_print_stack
The grep for loc_segfault_callback_print_stack was removed from the test
driver because it did not show up in the stacktrace output on
gabriel.sandia.gov.  Actually, I don't think this was supposed to show up in
the stacktrace after I changed the stacktrace code to not show stack tracing
functions by default.