Added --embed-rpath and --extra-configure-options
Added option --embed-rpath: Updates the specs file to put in rpath.  This was
suggested by Kevin Long.  I have not been able to get this to work on my
machine yet but it is an optional option so we can work on this iteratively.

Added option --extra-configure-options: Allows you to augment the options
passed to GCC's configure script.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=82,notpassed=0 (1.12 min)
1) SERIAL_RELEASE => passed: passed=82,notpassed=0 (1.03 min)