Various modifications to PyTrilinos documentation
The big change is a comprehensive build_docs script which can perform
a limited build of PyTrilinos and then run doxygen and the genereated
python files.  Also included in this commit are a newly made PDF of
the DevelopersGuide, updated Teuchos docstrings for SWIG and a new
Makefile for the Users Guide.