Added option for ignoring val() computations in Fad operators.
I added an option to allow for not computing the "value" in AD
operations when using the various Fad operators.  Clearly this only
works in some very special cases (such as linear problems), but in those
cases this can significantly reduce the cost of the Jacobian calculation
(particularly when PCE is embedded in the Jacobian calculation).  I also
migrated a change from CacheFad into the regular Fad to eliminate flops
in the * operation when one argument is a Fad, but has no derivative
components (again this is important when the underlying scalar type is a
PCE).  This required changing the expected number of flops in the scalar
flop counting example.