Changing to read CTEST_DROP_[METHOD,SITE,LOCATION] from env right in main file.
Here, I have put in a call to SET_DEFAULT_AND_FROM_ENV(...) to set the
varibles related to the CDash server right in the main
CTestConfig.cmake file.  This allows me to remove messy logic from
TrilinosCTestDriverCore.cmake.  The version of the code I committed
before was broken with respect to the drop site.

I have tested this manually to submit to a different CDash site and it
worked.

Build/Test Cases Summary
Enabled Packages: Teuchos, TrilinosFramework
0) MPI_DEBUG => passed: passed=153,notpassed=0 (0.98 min)
1) SERIAL_RELEASE => passed: passed=150,notpassed=0 (0.89 min)