Refactor epetra testing helper files.
bug# 5079

These files were being put into the base epetra library when testing mode was
on. This was a problem for the installation testing because then the installed
version of Trilinos that was built against had to have testing turned on which
doesn't make any sense. I decided that pulling out the files and putting them in
their own library made the most sense. This requires changing a few of the tests
to add the test only library to their dependencies.