Adding CMake utility script to create XML dependencies file (bug 4894)
Here I have create a new CMake script called
PackageArchDumpDepsXmlScript.cmake which is run with "cmake -P ...".  This is
used to generate a TilinosPackageDependencies.xml file which can include
packages in extra repos.  This will be needed to allow the checkin-test.py
script to be able to process with extra repos and add-on packages correctly.

I have also added some pretty good unit tests to show that this script is
working correctly and also shows how to run it.