Adding logic to enable a package if a file is deleted
Amazingly, I overlooked that if you delete a file from a package that the
checkin-test.py script will not enable the package to test the impact of
deleting that file!  I came across this recently when I deleted a single file
from MOOCHO:

   packages/moocho/src/MoochoUtils/src/FSeconds.f

which had been globbed but was not used anywhere but the checkin-test.py
script did not enable the MOOCHO package.

This commit adds logic and a unit test to prove that deleted files are flagged
and result in enabling the affected package.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=82,notpassed=0 (1.01 min)
1) SERIAL_RELEASE => passed: passed=82,notpassed=0 (0.90 min)