Added support for anisotropic bases.
The interface for CompletePolynomialBasis always supported an anisotropic
specification of the basis orders in each dimension, however the multi
to single index mapping and triple product generation didn't work if all
bases didn't have the same order.  This commit generalizes this code to
work in the anisotropic case as well.  I modified one stokhos example to
use such a basis, and it appears to work.