Add depends option to TRILINOS_DRIVER_ADD_DASHBOARD
Added a depends option that can be passed through to the test that is added by
this function. This will allow dashboards to specify any dependencies that it
has on other dashboards or tests. This was needed for installation testing so
that the installation build would not be run before or concurrently with the
build that creates the installation that will be used.