MueLu Aggregation: implemented option SetOrdering(1) (== random reorder)
The algorithm to set the seed before calling vector->randomize() was witchcraft. I kept the same kind of black magic but translated the jinx to the Teuchos dialect. My guess is that someone wrote this code to avoid some problem in ML with random generator. So I kept something similar...

Some remarks:
- The seed is compute with Teuchos for both Epetra and Tpetra.
  -> For Epetra, you cannot expect to get the same result as before because the seed will be different.
- randomize() use the random generator of Epetra for Epetra::Vector() and the generator of Teuchos (==std::rand()) for Tpetra::Vector()
  -> you cannot expect the same result when switching from Epetra to Tpetra. I don't see that as a problem.
- Added setSeed() to cthulhu vector
  -> this method doesn't exist in Tpetra but in Epetra, every vector have a random generator embedded and so I need this method to change the seed.