Fixing static validator problem; other cleanup
The main thing I fixed here was a problem where the LOWSF and PF validators
were static.  This did not allow you to create DefaultLinearSolverBuilder
objects and set different lists of LOWSF or PF factory objects on different
DLSB objects.  I moved these from static to be member objects.  I am not sure
why I did it this way when I first wrote this class but this was a major
implementation defect.  Eric Cyr discovered this when working on Teko where he
was adding custom Teko PF factory objects.

I have also added functions setDefaultLinearSolveStrategyFactoryName(...) and
setDefaultPreconditioningStrategyFactoryName(...), added makeDefault argument
to setLinearSolveStrategyFactory(...) and
setPreconditioningStrategyFactory(...).  This changes the behavior to not
change the default LOWSF or PF by default.  However, this still allows you to
set the LOWSF or PF as the default when calling these functions.  This is much
more clear.

I also added the function justInTimeInitialize() to make it a little more
clean that the just-in-time initialization idiom is being used.  This behavior
is a lot more clean now.

I also added a call to validateParameters(...) when calling
readParamaters(...) (which I don't think anyone actually uses).

Build/Test Cases Summary
Enabled Packages: Stratimikos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=251,notpassed=0 (26.76 min)
1) SERIAL_RELEASE => passed: passed=254,notpassed=0 (9.82 min)