Changing GCC builds to use MKL and removing need for env vars; adding boost to SS GCC builds
Here I have used CMake varible files to specify MKL and Boost
consistently across the various GCC and Intel builds.  MKL is just
BLAS/LAPACK so it is set for Primary Stable (PS) and Secondary Stable
(SS) builds of Trilinos.  However, Boost can only be set for SS builds
so Boost is added for GCC for only the nightly SS builds.

Here I have used these various CMake files to avoid duplication of
options for different builds.  I am not sure this is always the best
approach but I think it makes sense for CASL.  This ensures as much
consistency as possible between the nightly builds and the pre-push
builds with the checkin-test-fissile.sh script.