Adding dummy email address to allow to pass for user without git set up
It was discovered that a bunch of checkin-test.py unit tests were
failing because the tests assumed that the default --send-email-to
would always be non-empty.  This is not true if you don't have your
git email set.  This was the case for the special account
casl-vri-admin.  We have seen this error on other test machines but I
was never able to reproduce it because I alway had my git email set!

This commit adds a dummy email address for --send-email-to and it
results in all of these tests passing again.

I finally fixed this mysterious failure!

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=86,notpassed=0 (1.04 min)
1) SERIAL_RELEASE => passed: passed=86,notpassed=0 (0.84 min)