Fix configure breakage due to latest iRel API changes.
Mesquite previously checked for ${NAME}_dtor where ${NAME} is any of
iMesh, iGeom, or iRel to verify that the necessary library is present.
Apparently iRel_dtor was renamed iRel_destroy.  So check for either
${NAME)_dtor or ${NAME}_destroy because a) that way we're backwards
comaptible with older versions of iRel.h and b) we'll be forwards
compatible with the presumably upcomming analogous changes for iMesh
and iGeom.