Disable extra packages on windows shared build
There are a number of packages that have been explicitly enabled for the shared
build on windows which have not been set up to work properly on a shared build
for windows. I am disabling them so that we will only test the few packages that
have actually been setup to work properly in a shared build on windows.