Adding printing of TPL_TPLNAME_INCLUDE_DIRS and TPL_TPLNAME_LIBRARIES always
I did this so that it would be clear from the output exactly what TPL
libraries and include dirs are going to be used in the system without having
to grep the CMakeCache.txt file.  This does not greatly increase the output
volume and makes the output more explicit.  This was requested by Roger
Pawlowski.