Added ifdef protection around compareBeforeAndAfter methods
In Isorropia::Epetra::CostDescriber.cpp, the implementations for the
compareBeforeAndAfter methods are protected with an ifdef USE_UTILS;
in the header file they are not.  This creates problems for
PyTrilinos, because SWIG sees the methods in the header, generates
wrappers for them, and tries to call them -- but they are not in the
library.  By adding the same ifdefs to the header file, SWIG and thus
PyTrilinos, will handle them correctly.