Adding beginnings of files to drive a CI process for CASL VRI related software
Nothing is hooked up or tested but this is a reasonble start and I
need to commit and push because I am in the Dashboards/Trilinos dir
that will get rest by nightly testing soon.

Once I have this build working this will be the template that needs to
be used to update the CI testing on trilinos-test2 for Trilinos
proper.